+++ src/state/states.ts
@@ -13,6 +13,7 @@ import './game-model-state'
 import './pvp'
 import './hit-number'
 import './player-location'
+import './areas'
 
 declare global {
     namespace ig {